Move __version__ to version.py
